pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "t2lead"
version = "0.1.0"
description = "T2Lead: End-to-end drug discovery pipeline (Target → Hit → Lead)"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
dependencies = [
    "numpy",
    "pandas",
    "scikit-learn",
    "requests",
    "pyyaml",
    "python-dotenv",
]

[project.optional-dependencies]
dl = ["torch"]
h2l = ["crem"]
all = ["torch", "crem"]

[project.scripts]
t2lead = "drugpipe.pipeline:main"

[tool.setuptools.packages.find]
where = ["src"]
